Last Bit Of Life

+++ Team Project/FPS - 2507/Assets/Scripts/IDamage.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IDamage
 {
-    void takeDamage(int amount);
+    void takeDamage(float amount);
 }
