+++ Team Project/FPS - 2507/Assets/Scripts/bossAI.cs
@@ -19,7 +19,7 @@ public class bossAI : MonoBehaviour, IDamage
     [SerializeField] ParticleSystem deathAnim;
 
     [SerializeField] int goldDropped;
-    [SerializeField] int HP;
+    [SerializeField] float HP;
     [SerializeField] int fov;
     [SerializeField] int viewDistance;
     [SerializeField] int faceTargetSpeed;
@@ -286,7 +286,7 @@ public class bossAI : MonoBehaviour, IDamage
             agent.stoppingDistance = 0;
         }
     }
-    public void takeDamage(int amount)
+    public void takeDamage(float amount)
     {
         HP -= amount;
         agent.SetDestination(gameManager.instance.player.transform.position);
